requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.31.0
Pillow==10.0.1
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
gunicorn==21.2.0